+++ packages/react-drop-indicator/src/presets.tsx
@@ -1,12 +1,10 @@
 import { token } from '@atlaskit/tokens';
 
-import type { StrokeColor, StrokeWidth } from './internal-types';
+import type { Appearance, CSSColor } from './internal-types';
 
-export const presetStrokeColors: { [TKey in StrokeColor]: string } = {
-	standard: token('color.border.selected'),
+export const presetStrokeColors: { [TKey in Appearance]: CSSColor } = {
+	default: token('color.border.selected'),
 	warning: token('color.border.warning'),
 };
 
-export const presetStrokeSizes: { [TKey in StrokeWidth]: string } = {
-	standard: token('border.width.outline'),
-};
+export const presetStrokeWidth = token('border.width.outline');